simulazione3/challs/CTFe/CTFe/src/requirements.txt
Bootstrap-Flask==1.7.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
SQLAlchemy==1.4.20
supervisor==4.1.0
uWSGI==2.0.19.1
Werkzeug==2.0.1
WTForms==2.3.3